transfer_app/serializers.py
from rest_framework import serializers

from transfer_app.models import Transfer


class TransferSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transfer
        fields = '__all__'
