requirements.txt
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
executing==2.0.1
flake8==7.1.1
getenv==0.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
ipython==8.25.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
selenium==4.21.0
sentry-sdk==2.10.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions~=4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
utils==1.0.2
wcwidth==0.2.13
wsproto==1.2.0